rocr: Prevent int overflow in arithmetic operation
Cast range->x and range->y to uint64_t before performing multiplication



[ROCm/ROCR-Runtime commit: 77b86ca9084e5eccd9f2fa5d08fb61dec256609d]

+++ runtime/hsa-runtime/core/runtime/amd_blit_sdma.cpp
@@ -646,7 +646,7 @@ BlitSdma<RingIndexTy, HwIndexMonotonic, SizeToCountOffset, useGCR>::SubmitCopyRe
     BuildCopyRectCommand(append, dst, dst_offset, src, src_offset, range);
   }
 
-  uint64_t size = range->x * range->y * range->z;
+  uint64_t size = static_cast<uint64_t>(range->x) * static_cast<uint64_t>(range->y) * range->z;
 
   std::vector<core::Signal*> gang_signals(0);
